+++ LibrarySystem/LibrarySystem/Data/Repositories/AuthorRepository.cs
@@ -1,7 +1,7 @@
 ﻿using LibrarySystem.Models;
 using Microsoft.EntityFrameworkCore;
 
-namespace LibrarySystem.Data;
+namespace LibrarySystem.Data.Repositories;
 
 public class AuthorRepository(AppDBContext context) : BaseRepository<Author>(context), IAuthorRepository
 {
