+++ LibrarySystem/LibrarySystem/Data/Repositories/MessageRequestRepository.cs
@@ -1,6 +1,6 @@
 ﻿using LibrarySystem.Models;
 
-namespace LibrarySystem.Data;
+namespace LibrarySystem.Data.Repositories;
 
 public class MessageRequestRepository(AppDBContext dbcontext) 
     : BaseRepository<MessageRequest>(dbcontext),
